policies created

+++ app/Http/Controllers/CourseController.php
@@ -8,6 +8,7 @@ use App\Program;
 use App\Semester;
 use App\Course;
 use Yajra\DataTables\DataTables;
+use Auth;
 
 class CourseController extends Controller
 {
@@ -25,17 +26,19 @@ class CourseController extends Controller
             if($request->semester_id) {
                 $courses->where('semester_id', $request->semester_id);
             }
-            return $this->getDatatableOf($courses);
+            return $this->getDatatableOf($courses, Auth::user());
         }
         return view('course.index');
     }
 
-    private function getDatatableOf($courses) {
+    private function getDatatableOf($courses, $user) {
         return DataTables::of($courses)
         // ->editColumn('description', function($course){
         //     return Str::limit(strip_tags($course->description), 50);
         // })
-        ->addColumn('actions', function ($course) {
+        ->addColumn('actions', function ($course) use ($user) {
+            $actions['authorizedToEdit'] = $user->can('update', [Course::class, $course]);
+            $actions['authorizedToDelete'] = $user->can('delete', [Course::class, $course]);
             $actions['edit'] = route('course.edit', $course->id);
             $actions['delete'] = route('course.destroy', $course->id);
             return $actions;
@@ -50,6 +53,7 @@ class CourseController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Course::class);
         $programs = Program::select('id', 'name', 'abbreviation')->get();
         $semesters = Semester::select('id', 'name')->get();
         return view('course.create', ['programs'=>$programs, 'semesters'=>$semesters]);
@@ -63,17 +67,18 @@ class CourseController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Course::class);
         Course::create($request->all());
-        return redirect('/course');
+        return redirect('/course')->with(['message' => 'Course created sucessfully', 'alert-type' => 'success']);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  Course $course
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Course $course)
     {
         //
     }
@@ -81,14 +86,14 @@ class CourseController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  Course $course
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Course $course)
     {
+        $this->authorize('update', $course);
         $programs = Program::select('id', 'name', 'abbreviation')->get();
         $semesters = Semester::select('id', 'name')->get();
-        $course = Course::find($id);
         return view('course.create', [
             'course' => $course,
             'programs' => $programs,
@@ -100,26 +105,26 @@ class CourseController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  Course $course
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Course $course)
     {
-        $course = Course::find($id);
+        $this->authorize('update', $course);
         $course->update($request->all());
-        return redirect('/course');
+        return redirect('/course')->with(['message' => 'Course updated sucessfully', 'alert-type' => 'success']);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  Course $course
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Course $course)
     {
-        $course = Course::find($id);
+        $this->authorize('delete', $course);
         $course->delete();
-        return back();
+        return back()->with(['message' => 'Course deleted sucessfully', 'alert-type' => 'success']);
     }
 }
