+++ app/Http/Controllers/ProgramController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Program;
 use Yajra\DataTables\DataTables;
+use Auth;
 
 class ProgramController extends Controller
 {
@@ -17,14 +18,16 @@ class ProgramController extends Controller
     {
         if($request->ajax()) {
             $programs = Program::select();
-            return $this->getDatatableOf($programs);
+            return $this->getDatatableOf($programs, Auth::user());
         }
         return view('program.index'); 
     }
 
-    private function getDatatableOf($programs) {
+    private function getDatatableOf($programs, $user) {
         return DataTables::of($programs)
-        ->addColumn('actions', function ($program) {
+        ->addColumn('actions', function ($program) use ($user) {
+            $actions['authorizedToEdit'] = $user->can('update', [Program::class, $program]);
+            $actions['authorizedToDelete'] = $user->can('delete', [Program::class, $program]);
             $actions['edit'] = route('program.edit', $program->id);
             $actions['delete'] = route('program.destroy', $program->id);
             return $actions;
@@ -50,14 +53,15 @@ class ProgramController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Program::class);
         Program::create($request->all());
-        return redirect('/program');
+        return redirect('/program')->with(['message' => 'Program created sucessfully', 'alert-type' => 'success']);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  Program $program
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -68,12 +72,12 @@ class ProgramController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  Program $program
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(Program $program)
     {
-        $program = Program::find($id);
+        $this->authorize('update', $program);
         return view('program.create')->with('program', $program);
     }
 
@@ -81,27 +85,26 @@ class ProgramController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  Program $program
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Program $program)
     {
-        $program = Program::find($id);
         $program->update($request->all());
-        return redirect('/program');
+        return redirect('/program')->with(['message' => 'Program updated sucessfully', 'alert-type' => 'success']);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  Program $program
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Program $program)
     {
-        $program = Program::find($id);
+        $this->authorize('update', $program);
         $program->delete();
-        return back();
+        return back()->with(['message' => 'Program deleted sucessfully', 'alert-type' => 'success']);
 
     }
 }
